feat: split-bootstrap-portability-functionallong-helpers-001 - Isolate long-only portability helpers into explicit functionallong helper ownership
