+++ tests/functional/bootstrap_portability/cli_relative_working_directory_functionallong_helpers_test.go
@@ -1,3 +1,5 @@
+//go:build functionallong
+
 package bootstrap_portability
 
 import (
