+++ tests/functional/bootstrap_portability/helpers_http_test.go
@@ -1,0 +1,27 @@
+package bootstrap_portability
+
+import (
+	"encoding/json"
+	"net/http"
+	"testing"
+)
+
+func getGeneratedJSON[T any](t *testing.T, endpoint string) T {
+	t.Helper()
+
+	resp, err := http.Get(endpoint)
+	if err != nil {
+		t.Fatalf("GET %s: %v", endpoint, err)
+	}
+	defer resp.Body.Close()
+
+	if resp.StatusCode != http.StatusOK {
+		t.Fatalf("GET %s status = %d, want 200", endpoint, resp.StatusCode)
+	}
+
+	var out T
+	if err := json.NewDecoder(resp.Body).Decode(&out); err != nil {
+		t.Fatalf("decode %s: %v", endpoint, err)
+	}
+	return out
+}
