+++ tests/functional/bootstrap_portability/portability_helpers_test.go
@@ -1,8 +1,6 @@
 package bootstrap_portability
 
 import (
-	"os"
-	"path/filepath"
 	"testing"
 
 	"github.com/portpowered/infinite-you/pkg/petri"
@@ -21,21 +19,4 @@ func assertTokenPayload(t *testing.T, snap *petri.MarkingSnapshot, placeID, want
 	}
 
 	t.Fatalf("no token found in %s", placeID)
-}
-
-func writeFatFactoryJSON(t *testing.T, dir, content string) {
-	t.Helper()
-	if err := os.WriteFile(filepath.Join(dir, "factory.json"), []byte(content), 0o644); err != nil {
-		t.Fatalf("write fat factory.json: %v", err)
-	}
-}
-
-func writeFactoryTestFile(t *testing.T, path, content string) {
-	t.Helper()
-	if err := os.MkdirAll(filepath.Dir(path), 0o755); err != nil {
-		t.Fatalf("create parent dir for %s: %v", path, err)
-	}
-	if err := os.WriteFile(path, []byte(content), 0o644); err != nil {
-		t.Fatalf("write %s: %v", path, err)
-	}
 }
